Move sorting function improvements to a different file
